requirements.txt
langchain
langchain_openai
langchain_core
python_dotenv
langserve
fastapi
uvicorn
streamlit
langchain_community
sse_starlette
bs4
pypdf
faiss-cpu
langchain-text-splitters